♻️ refactor(telegram): Split telegram.go into focused files
Split the 853-line telegram.go into 5 files by responsibility:
- telegram.go (235): Bot struct, Config, New, Start, Notify, guard, group helpers
- handler.go (165): registerHandlers, handleText, command handlers
- model.go (146): type re-exports, model keyboard, pagination, switchModel
- stream.go (259): streaming (draft + edit), tool display, keepTyping
- render.go (86): markdown rendering, message splitting, sendChunkedMarkdown
